Support for UnsignedInt & UnsignedBigInt (#121)
* Initial unsigned support
* Fix to new nom-sql API
* Updated nom-sql to use the released v0.0.8
* Added arithmetic operations on unsigned ints

Thanks @alexsnaps!

+++ noria-server/dataflow/src/node/special/base.rs
@@ -240,8 +240,8 @@ impl Base {
                 match op {
                     Modification::Set(v) => future[col] = v,
                     Modification::Apply(op, v) => {
-                        let old: i64 = future[col].clone().into();
-                        let delta: i64 = v.into();
+                        let old: i128 = future[col].clone().into();
+                        let delta: i128 = v.into();
                         future[col] = match op {
                             Operation::Add => (old + delta).into(),
                             Operation::Sub => (old - delta).into(),
